+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -966,6 +966,9 @@
     "name": "createContainerRef"
   },
   {
+    "name": "createContentQuery"
+  },
+  {
     "name": "createElementNode"
   },
   {
@@ -2059,9 +2062,6 @@
   },
   {
     "name": "ɵɵattribute"
-  },
-  {
-    "name": "ɵɵcontentQuery"
   },
   {
     "name": "ɵɵdefineComponent"
